refactor(console): extract format helpers to lib/format.ts
Move formatDuration / formatRelative / shortenId / pickTickStep out
of SessionDetail.tsx (1833 → 1794 lines, headed lower) into a
dedicated lib/format.ts so they can be reused from other pages and
tested independently. No behaviour change.
